refactor skr package paths
