+++ components/kcp/internal/controller/cloud-resources/awsnfsvolume_controller.go
@@ -18,7 +18,7 @@ package cloudresources
 
 import (
 	"context"
-	"github.com/kyma-project/cloud-manager/components/kcp/pkg/skr/registry"
+	skrruntime "github.com/kyma-project/cloud-manager/components/kcp/pkg/skr/runtime"
 	"k8s.io/klog/v2"
 	"sigs.k8s.io/controller-runtime/pkg/cluster"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
@@ -67,7 +67,7 @@ func (r *AwsNfsVolumeReconciler) Reconcile(ctx context.Context, req ctrl.Request
 	return ctrl.Result{}, nil
 }
 
-func SetupAwsNfsVolumeReconciler(reg registry.SkrRegistry) error {
+func SetupAwsNfsVolumeReconciler(reg skrruntime.SkrRegistry) error {
 	return reg.Register().
 		WithFactory(&AwsNfsVolumeReconcilerFactory{}).
 		For(&cloudresourcesv1beta1.AwsNfsVolume{}).
