+++ components/kcp/internal/controller/cloud-resources/iprange_controller.go
@@ -19,7 +19,7 @@ package cloudresources
 import (
 	"context"
 	cloudresourcesv1beta1 "github.com/kyma-project/cloud-manager/components/kcp/api/cloud-resources/v1beta1"
-	"github.com/kyma-project/cloud-manager/components/kcp/pkg/skr/registry"
+	skrruntime "github.com/kyma-project/cloud-manager/components/kcp/pkg/skr/runtime"
 	"k8s.io/klog/v2"
 	"sigs.k8s.io/controller-runtime/pkg/cluster"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
@@ -66,7 +66,7 @@ func (r *IpRangeReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ct
 	return ctrl.Result{}, nil
 }
 
-func SetupIpRangeReconciler(reg registry.SkrRegistry) error {
+func SetupIpRangeReconciler(reg skrruntime.SkrRegistry) error {
 	return reg.Register().
 		WithFactory(&IpRangeReconcilerFactory{}).
 		For(&cloudresourcesv1beta1.CloudResources{}).
