+++ components/kcp/pkg/skr/runtime/alias.go
@@ -1,0 +1,12 @@
+package runtime
+
+import (
+	"github.com/kyma-project/cloud-manager/components/kcp/pkg/skr/runtime/looper"
+	"github.com/kyma-project/cloud-manager/components/kcp/pkg/skr/runtime/registry"
+)
+
+var NewLooper = looper.New
+
+var NewRegistry = registry.New
+
+type SkrRegistry = registry.SkrRegistry
