+++ components/kcp/pkg/skr/runtime/looper/skrLooper.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 	"github.com/elliotchance/pie/v2"
 	"github.com/go-logr/logr"
-	skrmanager "github.com/kyma-project/cloud-manager/components/kcp/pkg/skr/manager"
-	"github.com/kyma-project/cloud-manager/components/kcp/pkg/skr/registry"
+	skrmanager "github.com/kyma-project/cloud-manager/components/kcp/pkg/skr/runtime/manager"
+	"github.com/kyma-project/cloud-manager/components/kcp/pkg/skr/runtime/registry"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/cluster"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
